Refactor server code by separating routers
